fix: clone template to tmp

+++ src/lib/fs_utils.cpp
@@ -95,15 +95,20 @@ void Templi::process_for_each_line(String path,
 	file.close();
 }
 
-void delete_folder(String path)
+void Templi::delete_folder(String path)
 {
-	try
+	if (!fs::exists(path))
+	{
+		return;
+	}
+	
+    try
 	{
 		fs::remove_all(path);
 	}
 	catch (const std::filesystem::filesystem_error &e)
 	{
 		String message = e.what();
-		throw new Templi::Exception("Cannot delete" + message);
+		throw Templi::Exception(message);
 	}
 }
